+++ CHANGES.rst
@@ -5,6 +5,7 @@ USPTO PAIR Bulk Data API Client CHANGES
 development
 ===========
 - Implement the basic query and download parts of the USPTO PAIR Bulk Data API
+- Implement Celery-based downloader infrastructure
 
 2017-10-19 0.0.0
 ================